feat: implement configurable API key storage
Introduce an APIKeyStorage interface with two concrete implementations:
KeyringAPIKeyStorage and FileAPIKeyStorage.

The FileAPIKeyStorage implementation is first of all intended for WSL2 environments
where keyring does not work that well, so temporarly storing to the file-system is
the most user-friendly option.

The ResolveAPIKeyStorage function resolves which storage mechanism to use. If the system
is WSL2, automatically use FileAPIKeyStorage.

Refactor config package to align with the new storage mechanism.